### 2021-11-25 #489 @helske
Submission of bssm for Bayesian state space modelling
Submitting Author: Jouni Helske (@helske)
Other Package Authors: (delete if none) Name (@mvihola)
Repository:  https://github.com/helske/bssm
Version submitted: 2.0.0
Submission type: Stats
Badge grade: silver

Version accepted: TBD
---

```
Package: bssm
Title: Bayesian Inference of Non-Linear and Non-Gaussian State Space
        Models
Version: 2.0.0
    c(person(given = "Jouni",
           family = "Helske",
           email = "jouni.helske@iki.fi",
           comment = c(ORCID = "0000-0001-7130-793X")),
      person(given = "Matti",
           family = "Vihola",
           role = "aut",
           comment = c(ORCID = "0000-0002-8041-7222")))
Description: Efficient methods for Bayesian inference of state space models
    via particle Markov chain Monte Carlo (MCMC) and MCMC based on parallel
    importance sampling type weighted estimators
    (Vihola, Helske, and Franks, 2020, <doi:10.1111/sjos.12492>).
    Gaussian, Poisson, binomial, negative binomial, and Gamma
    observation densities and basic stochastic volatility models
    with linear-Gaussian state dynamics,
    as well as general non-linear Gaussian models and discretised
    diffusion models are supported.
    covr,
    ggplot2 (>= 2.0.0),
    KFAS (>= 1.2.1),
    knitr (>= 1.11),
    MASS,
    rmarkdown (>= 0.8.1),
    ramcmc,
    sde,
    sitmo,
    checkmate,
    coda (>= 0.18-1),
    diagis,
    posterior,
    Rcpp (>= 0.12.3),
    tidyr
LinkingTo: ramcmc, Rcpp, RcppArmadillo, sitmo
SystemRequirements: C++11, pandoc (>= 1.12.3, needed for vignettes)
NeedsCompilation: yes
Roxygen: list(markdown = TRUE,
  roclets = c("namespace", "rd", "srr::srr_stats_roclet"))

```

- [ ] A pre-submission inquiry has been approved in [issue#<issue_num>](#<issue_num>)
I have not made a pre-submission inquiry, but was asked to consider submitting by @mpadge and @noamross.

State space models provide a flexible framework for statistical inference of a broad class of time series and other dynamic data. The `bssm` package aims to provide easy to use and efficient functions for the Bayesian estimation of commonly used as well as more general user-defined state space models, which are usable in various application areas.

This is the first software to implement the IS-MCMC by
Vihola, Helske, and Franks (2020) and first R package to implement delayed
acceptance pseudo-marginal MCMC for state space models. The IS-MCMC method
is also available in [walker](github.com/helske/walker) package for a
limited class of time-varying GLMss (a small subset of the models
supported by this package). Some of the functionality for exponential family
state space models is also available in [KFAS](github.com/helske/KFAS), and
those models can be converted easily to bssm format for Bayesian analysis.

Silver sounds appropriate.

The `bssm` complies with a large number of standards both in the general category as well as in Bayesian Software category and their sub-categories. I see the package complying with several Time Series Software standards as well, although many of those standards do not seem to be well suited or applicable to general time series modelling via state space models and/or `bssm`, so at least for now I have focused on the General and Bayesian standards.

The modelling framework and the implemented algorithms are very general, and since the early versions, the usability and features of the `bssm` are greatly improved to quite general models and applications (Currently `bssm` has most of the same features and many more as in the popular `KFAS` package for state space modelling which has been used in various domains).

- [x] I/we have read the [guide for authors](https://devguide.ropensci.org/guide-for-authors.html) and [rOpenSci packaging guide](https://devguide.ropensci.org/building.html).
- [x] I/we have read the [*Statistical Software Peer Review* Guide for Authors](https://ropenscilabs.github.io/statistical-software-review-book/pkgdev.html). *(this link doesn't actually work)*
(there are few problems for which I have opened an issue/PR in the autotest repo).

This package is already on CRAN.
