### 2021-11-25 08ebfbf
2.0.0 to rOpensci and CRAN
