+++ README.Rmd
@@ -9,7 +9,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.path = "man/figures/README-",
-  out.width = "100%"
+  out.width = "100%",
+  cache = TRUE
 )
 ```
 
@@ -68,9 +69,9 @@ knitr::opts_chunk$set(
 <!-- badges: start -->
 [![Project Status: Active - The project has reached a stable, usable state and is being actively developed](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![R-CMD-check](https://github.com/helske/bssm/workflows/R-CMD-check/badge.svg)](https://github.com/helske/bssm/actions)
-[![Codecov test coverage](https://codecov.io/gh/helske/bssm/branch/master/graph/badge.svg)](https://codecov.io/gh/helske/bssm?branch=master)
-[![CRAN version](http://www.r-pkg.org/badges/version/bssm)](http://cran.r-project.org/package=bssm)
-[![downloads](http://cranlogs.r-pkg.org/badges/bssm)](http://cranlogs.r-pkg.org/badges/bssm)
+[![Codecov test coverage](https://codecov.io/gh/helske/bssm/branch/master/graph/badge.svg)](https://app.codecov.io/gh/helske/bssm?branch=master)
+[![CRAN version](http://www.r-pkg.org/badges/version/bssm)]( https://CRAN.R-project.org/package=bssm)
+[![downloads](https://cranlogs.r-pkg.org/badges/bssm)](https://cranlogs.r-pkg.org/badges/bssm)
 
 <!-- badges: end -->
 
@@ -85,7 +86,7 @@ supported.
 For details, see 
 
 * [The bssm paper on ArXiv](https://arxiv.org/abs/2101.08492) (to appear in R Journal), 
-* [Package vignettes at CRAN](https://cran.r-project.org/web/packages/bssm/index.html) 
+* [Package vignettes at CRAN](https://CRAN.R-project.org/package=bssm) 
 * Paper on [Importance sampling type estimators based on approximate marginal Markov chain Monte Carlo](https://onlinelibrary.wiley.com/doi/abs/10.1111/sjos.12492)
 
 There are also couple posters and a talk related to IS-correction methodology and bssm package: 
