### 2021-11-26 d4f5c9f
remove nonexistent srr tag, fixes #30#
