### 2021-12-10 b9ccc0d
set weights to 1
