### 2022-05-02 16f619c
tweak tolerances to pass MKL checks on CRAN, update citation info
