+++ tests/testthat/test_basics.R
@@ -2,6 +2,8 @@ context("Test basics")
 
 #' @srrstats {G5.4, G5.4b, G5.6, G5.6a, G5.6b, G5.7} Compare with KFAS.
 
+tol <- 1e-6
+
 test_that("results for Gaussian models are comparable to KFAS", {
   library("KFAS")
   model_KFAS <- SSModel(1:10 ~ SSMtrend(2, Q = list(0.01^2, 0)), H = 2)
@@ -14,15 +16,15 @@ test_that("results for Gaussian models are comparable to KFAS", {
     sd_y = 0.01))
   model_bssm <- bsm_lg(1:10, P1 = diag(1e2, 2), sd_slope = 0,
     sd_level = 0.01, sd_y = sqrt(2))
-  
+
   expect_equal(logLik(model_KFAS, convtol = 1e-12), logLik(model_bssm, 0))
   out_KFAS <- KFS(model_KFAS, filtering = "state", convtol = 1e-12)
   expect_error(out_bssm <- kfilter(model_bssm), NA)
-  expect_equivalent(out_KFAS$a, out_bssm$at)
-  expect_equivalent(out_KFAS$P, out_bssm$Pt)
+  expect_equivalent(out_KFAS$a, out_bssm$at, tolerance = tol)
+  expect_equivalent(out_KFAS$P, out_bssm$Pt, tolerance = tol)
   expect_error(out_bssm <- smoother(model_bssm), NA)
-  expect_equivalent(out_KFAS$alphahat, out_bssm$alphahat)
-  expect_equivalent(out_KFAS$V, out_bssm$Vt)
+  expect_equivalent(out_KFAS$alphahat, out_bssm$alphahat, tolerance = tol)
+  expect_equivalent(out_KFAS$V, out_bssm$Vt, tolerance = tol)
 })
 
 test_that("results for multivariate Gaussian model are comparable to KFAS", {
@@ -36,84 +38,85 @@ test_that("results for multivariate Gaussian model are comparable to KFAS", {
         Q = matrix(1), P1inf = diag(2)) +
       SSMseasonal(period = 12, sea.type = "trigonometric"),
     data = Seatbelts, H = matrix(NA, 2, 2))
-  
+
   diag(kfas_model$P1) <- 50
   diag(kfas_model$P1inf) <- 0
-  kfas_model$H <- structure(c(0.00544500509177812, 0.00437558178720609, 
+  kfas_model$H <- structure(c(0.00544500509177812, 0.00437558178720609,
     0.00437558178720609, 0.00885692410165593), .Dim = c(2L, 2L, 1L))
   kfas_model$R <- structure(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0.0152150188066314, 0.0144897116711475
-  ), .Dim = c(29L, 1L, 1L), .Dimnames = list(c("log(PetrolPrice).front", 
-    "log(kms).front", "log(PetrolPrice).rear", "log(kms).rear", "law.front", 
-    "sea_trig1.front", "sea_trig*1.front", "sea_trig2.front", 
-    "sea_trig*2.front", "sea_trig3.front", "sea_trig*3.front", 
-    "sea_trig4.front", "sea_trig*4.front", 
-    "sea_trig5.front", "sea_trig*5.front", "sea_trig6.front", "sea_trig1.rear", 
-    "sea_trig*1.rear", "sea_trig2.rear", "sea_trig*2.rear", "sea_trig3.rear", 
[…]
