+++ README.Rmd
@@ -85,7 +85,7 @@ supported.
 
 For details, see 
 
-* [The bssm paper on ArXiv](https://arxiv.org/abs/2101.08492) (to appear in R Journal), 
+* [The bssm paper on The R Journal](https://journal.r-project.org/archive/2021/RJ-2021-103/index.html), 
 * [Package vignettes at CRAN](https://CRAN.R-project.org/package=bssm) 
 * Paper on [Importance sampling type estimators based on approximate marginal Markov chain Monte Carlo](https://onlinelibrary.wiley.com/doi/abs/10.1111/sjos.12492)
 
@@ -95,15 +95,13 @@ There are also couple posters and a talk related to IS-correction methodology an
 * [SMC 2017 workshop: Accelerating MCMC with an approximation ](http://users.jyu.fi/~jovetale/posters/SMC2017)
 * [UseR!2017: Bayesian non-Gaussian state space models in R](http://users.jyu.fi/~jovetale/posters/user2017.pdf)
 
-The `bssm` package was originally developed with the support of Academy of 
-Finland grants 284513, 312605, and 311877. Current development is focused on 
-increased usability. For recent changes, see NEWS file.
+The `bssm` package was originally developed with the support of Academy of Finland grants 284513, 312605, 311877, and 331817. Current development is focused on increased usability. For recent changes, see NEWS file.
 
 ### Citing the package 
 
 If you use the `bssm` package in publications, please cite the corresponding R Journal paper:
 
-Jouni Helske and Matti Vihola (2021). "bssm: Bayesian Inference of Non-linear and Non-Gaussian State Space Models in R." _R Journal_. To appear. ArXiv preprint https://arxiv.org/abs/2101.08492.
+Jouni Helske and Matti Vihola (2021). "bssm: Bayesian Inference of Non-linear and Non-Gaussian State Space Models in R." The R Journal (2021) 13:2, pages 578-589. https://journal.r-project.org/archive/2021/RJ-2021-103/index.html
 
 ## Installation
 
