### 2022-05-06 #489 @maelle
@Athene-ai you're a reviewer for https://github.com/ropensci/software-review/issues/523 (thank you) so you can't be added as reviewer to other packages for a little while.

I realize our dev guide might lack guidance on how to volunteer as reviewer, so we'll add it https://github.com/ropensci/dev_guide/issues/457 Thank you!
