### 2022-05-06 745373d
add ropensci badge, switch to markdown NEWS
