+++ README.md
@@ -8,6 +8,8 @@
 [![Project Status: Active - The project has reached a stable, usable
 state and is being actively
 developed](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/489_status.svg)](https://github.com/ropensci/software-review/issues/489)
 [![R-CMD-check](https://github.com/helske/bssm/workflows/R-CMD-check/badge.svg)](https://github.com/helske/bssm/actions)
 [![Codecov test
 coverage](https://codecov.io/gh/helske/bssm/branch/master/graph/badge.svg)](https://app.codecov.io/gh/helske/bssm?branch=master)
@@ -26,7 +28,8 @@ models and discretely observed latent diffusion processes are supported.
 
 For details, see
 
--   [The bssm paper on R Journal](https://journal.r-project.org/archive/2021/RJ-2021-103/index.html),
+-   [The bssm paper on The R
+    Journal](https://journal.r-project.org/archive/2021/RJ-2021-103/index.html),
 -   [Package vignettes at CRAN](https://CRAN.R-project.org/package=bssm)
 -   Paper on [Importance sampling type estimators based on approximate
     marginal Markov chain Monte
@@ -43,8 +46,9 @@ methodology and bssm package:
     R](http://users.jyu.fi/~jovetale/posters/user2017.pdf)
 
 The `bssm` package was originally developed with the support of Academy
-of Finland grants 284513, 312605, 311877, and 331817. Current development is
-focused on increased usability. For recent changes, see NEWS file.
+of Finland grants 284513, 312605, 311877, and 331817. Current
+development is focused on increased usability. For recent changes, see
+NEWS file.
 
 ### Citing the package
 
@@ -52,7 +56,9 @@ If you use the `bssm` package in publications, please cite the
 corresponding R Journal paper:
 
 Jouni Helske and Matti Vihola (2021). “bssm: Bayesian Inference of
-Non-linear and Non-Gaussian State Space Models in R.” *R Journal*. https://journal.r-project.org/archive/2021/RJ-2021-103/index.html>.
+Non-linear and Non-Gaussian State Space Models in R.” The R Journal
+(2021) 13:2, pages 578-589.
+<https://journal.r-project.org/archive/2021/RJ-2021-103/index.html>
 
 ## Installation
 
@@ -88,12 +94,14 @@ allowed);
 
 ``` r
 library("bssm")
+#> Warning: package 'bssm' was built under R version 4.1.3
 #> 
 #> Attaching package: 'bssm'
 #> The following object is masked from 'package:base':
 #> 
 #>     gamma
 library("dplyr")
+#> Warning: package 'dplyr' was built under R version 4.1.3
 #> 
 #> Attaching package: 'dplyr'
 #> The following objects are masked from 'package:stats':
@@ -103,6 +111,7 @@ library("dplyr")
 #> 
 #>     intersect, setdiff, setequal, union
 library("ggplot2")
[…]
