+++ NEWS.md
@@ -1,3 +1,8 @@
+bssm 2.0.x (2022-)
+=====================================
+* Switched to markdown NEWS with a plan to be more clear about the future 
+  changes in the package.
+
 bssm 2.0.1 (Release date: 2022-05-02)
 ==============
    * Fixed weights to one in case of non-linear model with mcmc_type="approx".
