+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: bssm
 Type: Package
 Title: Bayesian Inference of Non-Linear and Non-Gaussian State Space
         Models
-Version: 2.0.1
+Version: 2.0.1.1
 Authors@R: 
     c(person(given = "Jouni",
            family = "Helske",
