### 2022-06-25 #489 @kingaa
The documentation of standards is admirable.

Examining the source code, it appears to be very carefully written and in a good style.  I expect it will be easy to track down bugs and to maintain the code written in this style.

Adopting the user's point of view, I attempted to plunge in. I found it more difficult to do so than I suspect the authors would like.  This leads to some suggestions regarding the documentation.

First, it would be good if `package?bssm` gave documentation on the package as a whole, with orientation toward the major features for the novice user.  The help available from `library(help=bssm)` is mnemonic but not a good introduction.

I attempted to follow the "bssm" vignette.  The first example, which begins
```
data("nhtemp", package = "datasets")
prior <- halfnormal(1, 10)
```
left me wondering about the "half-normal" prior.  The man page on priors is minimally informative.  I do not see formulae for the prior densities, nor are there descriptions of their parametrization, nor even plots.  The examples there seem mainly to be for automated-checking purposes: they shed little light for the user.  Improved documentation would be helpful, as would examples of their actual usage, and some plots.

Continuing with the example, I did
```
bsm_model <- bsm_lg(y=nhtemp,sd_y=prior,sd_level=prior,sd_slope=prior)
mcmc_bsm <- run_mcmc(bsm_model, iter = 4e4, seed = 1)
summary(mcmc_bsm)
mcmc_bsm
plot(mcmc_bsm)
```
This did give some information, though the `plot` method failed.  Following the vignette, I was able to plot the approximate posterior densities using
```
mcmc_bsm |>
  as.data.frame() |>
  ggplot(aes(x=value,group=variable))+
  geom_density()+
  facet_wrap(~variable,scales="free")
```
However, I imagine other experienced **R** users, like me, would appreciate more informative outputs from the `summary` and `plot` methods.

Also, I immediately found myself wanting to know:
- How can MCMC convergence be diagnosed?
- How can I make a plot showing prior and posterior, perhaps on the same axes?
- How I run multiple independent chains?

I notice that, among the list of standards deemed inapplicable by the package authors are some that speak to these questions, and to the issues with understanding prior distributions which I mentioned before.

I also noticed that this vignette mentions and discusses, but does not demonstrate, nonlinear, non-Gaussian models.  Since such models are a major feature of the package, some demonstration would be appreciated.

Turning to the "growth_model" vignette, I was intrigued to see that there are facilities for including snippets of C++ code.  However, I was not able to follow the vignette sufficiently well as to be able to reproduce those calculations myself.  I would appreciate more detailed, step-by-step instructions on how to compile the snippets shown.  (For example, I get errors regarding the unknown namespace `arma`).

Though there is more exploration I would like to do, I will stop here for now.  It is my understanding (though I am happy to be corrected) that this review process is intended to include back-and-forth.  Some comments from the authors may help me complete what I hope will be a useful review.
