### 2023-02-03 #489 @maurolepore
Sorry to jump in. My EiC rotation just started and I'm checking the status of every open issue.

I'm curious about a few things:

- [ ] 1. I see only one reviewer. @noamross are we we searching for a second reviewer or leave it at one?

- [ ] 2. @noamross I see you replaced @bbolker. Do we need to update the top comment with the latest editor?

- [ ] 3. @kingaa RE
> Though there is more exploration I would like to do, I will stop here for now. It is my understanding (though I am happy to be corrected) that this review process is intended to include back-and-forth.

Typically we ask reviewers to complete their review in one go. The authors address the comments of both reviewers. Then the reviewers either (a) approve the changes or (b) request more changes.

However, before you use more of your valuable time let's see what response I get.
