### 2024-11-28 #489 @helske
Hi, I haven't actively pushed this forward as I mostly work with different projects now, but it would be nice to have some kind of closure for this. I did already comment the first review earlier and made changes accordingly, and if you can find second reviewer, I'm happy to engage with the review again. I'm fine with other possible solutions as well.
