### 2025-02-02 #489 @maurolepore
Dear @helske this is to mark the start of my EiC rotation. I'm reviewing all open issues and noting what I see:

* @helske has been responsive and intends to continue.
* @bbolker was the original editor but @noamross took over.
* @kingaa is  one reviewer and shared some feedback but the formal review is pending.
* A second reviewer is still missing.

@helske I conclude the delays re not on your part. I'm sorry. Can you confirm you intend to continue? If so I'll remove the holding and help move this forward.
