### 2025-05-04 #489 @maurolepore
@helske I'm wrapping up my EiC rotation today. As a summary for the next EiC I hightlight these two comments:

* https://github.com/ropensci/software-review/issues/489#issuecomment-2656285094
* https://github.com/ropensci/software-review/issues/489#issuecomment-2629509020

Also:

* This is a stats submission,
* we're missing a second reviewer,
* it's a bit unclear who's acting as the handling editor
* it's tagged as holding
* as of mid Feb @helske intended to resume.

@helske we're sorry for the slow communication. I'll share this with our editorial board. It might help to once again that you're willing to resume and please let us know if you're getting notifications in a timely manner.
