### 2026-05-03 #489 @maurolepore
Dear @helske, this is to mark the start of my new EiC rotation.

Brief summary of what I see:
- Reviewer 1 (@kingaa) shared informal feedback long ago but never posted a formal review.
- Reviewer 2 was never found.
- You confirmed in February 2025 that you intend to continue.
- @mpadge generously committed in November to look for a second reviewer; no update since.

The intent on both sides points to removing the `holding` label. But the responsiveness on both sides suggests keeping it. That's the tension I'd like to resolve.

@helske, could you confirm once more whether you'd like us to keep moving? I'll also drop you a note on Slack since I recall GitHub mentions don't always reach you. Thanks for your patience.
