### 2026-05-04 #489 @helske
Thanks Mauro. I did indicate a year ago that I was willing to continue (IIRC, in email with Mark), but given the apparent challenges in finding reviewers, and more importantly the fact that I have moved on to other research topics and do not currently actively develop or use `bssm`, I think it makes sense to abandon this.
